Improve: Introduce a prelude for BC1 API testing
