+++ projects/api/dxt-lossless-transform-bc1-api/src/determine_optimal_transform.rs
@@ -126,29 +126,7 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use dxt_lossless_transform_api_common::estimate::{DataType, SizeEstimationOperations};
-
-    // Create a simple dummy estimator for testing
-    struct DummyEstimator;
-
-    impl SizeEstimationOperations for DummyEstimator {
-        type Error = &'static str;
-
-        fn max_compressed_size(&self, _len_bytes: usize) -> Result<usize, Self::Error> {
-            Ok(0) // No buffer needed for dummy estimator
-        }
-
-        unsafe fn estimate_compressed_size(
-            &self,
-            _input_ptr: *const u8,
-            len_bytes: usize,
-            _data_type: DataType,
-            _output_ptr: *mut u8,
-            _output_len: usize,
-        ) -> Result<usize, Self::Error> {
-            Ok(len_bytes) // Just return the input length
-        }
-    }
+    use crate::test_prelude::DummyEstimator;
 
     #[test]
     fn test_determine_optimal_transform_with_options() {
